% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{adaptive_centrality_influential}
\alias{adaptive_centrality_influential}
\title{Returns the most influential nodes in a graph using adaptive centrality-based heuristics}
\usage{
adaptive_centrality_influential(
  graph,
  budget = 1,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC"),
  centrality_method = c("DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}

\item{centrality_method}{defines the centrality method to be used. Value must be:}
}
\value{
object containing: 1. Vector of influential nodes. 2. Measure of influence. 3. Elapsed time in seconds.
}
\description{
Returns the most influential nodes in a graph using adaptive centrality-based heuristics
}
\references{
Lipton, R. J., & Naughton, J. F. (1989). Estimating the size of generalized transitive closures. In Proceedings of the 15th Int. Conf. on Very Large Data Bases.
}
