% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_mining_framework.R
\name{celf_greedy_framework}
\alias{celf_greedy_framework}
\title{Influence mining framework implementation based on CELF greedy method}
\usage{
celf_greedy_framework(
  graph,
  criteria = c("TIME", "BUDGET", "INFLUENCE"),
  threshold,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{criteria}{is the stopping criteria for influence framework. Supported values are TIME, BUDGET and SPREAD}

\item{threshold}{is the limit value for criteria function. The execution steps as soon as criteria meets the threshold value. If the criteria is TIME, then threshold should be defined in seconds}

\item{test_method}{specifies the method to measure influence. Value "RESILIENCE" (number of total nodes REMOVED (NOT THE REMAINING ones as in original resilience function) from the graph); "INFLUENCE_IC" (see simulate_ic method); "INFLUENCE_LT" (see simulate_lt method). Default is "RESILIENCE"}
}
\value{
output object containing influential nodes, time taken and the influence of nodes according the test_method
}
\description{
Influence mining framework implementation based on CELF greedy method
}
