% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{celf_plus_plus}
\alias{celf_plus_plus}
\title{Returns the set of influential nodes identified by CELF++ algorithm}
\usage{
celf_plus_plus(graph, budget = 1, test_method)
}
\arguments{
\item{graph}{the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}
}
\value{
output containing summary
}
\description{
Returns the set of influential nodes identified by CELF++ algorithm
}
\examples{
{celf_plus_plus(graph=erdos.renyi.game(100, 0.2), 10, "INFLUENCE_LT")}
}
\references{
Leskovec, J., Krause, A., Guestrin, C., Faloutsos, C., VanBriesen, J., & Glance, N. (2007). Cost-effective Outbreak Detection in Networks. Proceedings of the 13th ACM SIGKDD International Conference on Knowledge Discovery and Data Mining - KDD ’07, 420–429.
}
