% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{collective_influence}
\alias{collective_influence}
\title{Returns CI value of given graph and node}
\usage{
collective_influence(
  graph,
  neighborhood_distance,
  node_id,
  method = c("degree")
)
}
\arguments{
\item{graph}{the igraph object}

\item{neighborhood_distance}{is the distance to which the neighborhood nodes are searched for}

\item{node_id}{is the ID of the target node}

\item{method}{is the metric to calculate sum of influence. Default is "degree"}
}
\value{
influence as product of degree of target node and total sum of degrees of neighborhood
TODO: extend the function and include adaptive methods as well as other centrality methods
}
\description{
Returns CI value of given graph and node
}
\examples{
{collective_influence(graph=erdos.renyi.game(100, 0.2), neighborhood_distance=2, 1)}
}
