% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{community_modularity_test}
\alias{community_modularity_test}
\title{Detects function performs modularity test of communities.}
\usage{
community_modularity_test(graph)
}
\arguments{
\item{graph}{is igraph object}
}
\description{
Detects function performs modularity test of communities.
}
