% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{community_significance_test}
\alias{community_significance_test}
\title{Detects function performs a Wilcoxon rank-sum test on the "internal" and "external" degrees of a community in order to quantify its significance.}
\usage{
community_significance_test(graph, vs, ...)
}
\arguments{
\item{graph}{is igraph object}

\item{vs}{is the vector of vertices of the original graph which will form the subgraph}

\item{...}{arguments to pass to \link{wilcox.test} function}
}
\description{
The edges within a community are "internal" and the edges connecting the vertices of a community with the rest of the graph are "external". More internal than external edges show that the community is significant; the otherwise suggests that the community is in fact an "anti-community".
}
