% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_util.R
\name{discretize_naturally}
\alias{discretize_naturally}
\title{Discretizes a vector naturally}
\usage{
discretize_naturally(x, breaks = 7)
}
\arguments{
\item{x}{vector to discretize}

\item{breaks}{is the number of classes. This can only be 3, 5 or 7}
}
\value{
descritized vector
}
\description{
Discretizes a numeric vector into 3, 5 or 7 and returns a factor vector with values as: LOW, MEDIUM, HIGH; VERY_LOW, LOW, MEDIUM, HIGH, VERY_HIGH; VERY_LOW, LOW, MEDIUM_LOW, MEDIUM, MEDIUM_HIGH, HIGH, VERY_HIGH
}
