% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{find_communities}
\alias{find_communities}
\title{Finds communities in the given graph and returns the graph after adding a vector "group" to its vertices}
\usage{
find_communities(
  graph,
  plot = TRUE,
  method = c("multilevel", "edgebetweenness", "fastgreedy", "eigenvector", "spinglass",
    "walktrap", "labelpropagation", "clique", "largescale")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{plot}{whether to plot the \code{graph} using \link{tkplot}. Default is TRUE}

\item{method}{is the method to find communities. Value can be "multilevel", "edgebetweenness", "fastgreedy", "eigenvector", "spinglass", "walktrap", "labelpropagation", "clique", "largescale"}
}
\value{
\code{graph} object with membership attribute attached representing communities
}
\description{
Finds communities in the given graph and returns the graph after adding a vector "group" to its vertices
}
