% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{fit_power_law}
\alias{fit_power_law}
\title{Plots degree distribution and returns power-law exponent of given graph}
\usage{
fit_power_law(graph)
}
\arguments{
\item{graph}{is the igraph object}
}
\description{
Plots degree distribution and returns power-law exponent of given graph
}
