% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_clique}
\alias{generate_clique}
\title{Generates a fully connected undirected graph}
\usage{
generate_clique(size)
}
\arguments{
\item{size}{is the number of nodes}
}
\value{
igraph object
}
\description{
Generates a fully connected undirected graph
}
