% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_holme_kim}
\alias{generate_holme_kim}
\title{Generates a Holme-Kim Network}
\usage{
generate_holme_kim(size, m, triad_prob = 0.1, directed = FALSE)
}
\arguments{
\item{size}{is the number of nodes of the network}

\item{m}{is the number of nodes to which a new node connects at each iteration}

\item{triad_prob}{is Triad formation probability after each preferential attachment mechanism}

\item{directed}{whether the graph is directed or not. Default is FALSE}
}
\value{
A list containing the nodes of the network and their respective neighbors.
}
\description{
Simulate a scale-free network with relatively high clustering, comparing to B-A networks (Holme and Kim, 1999).
}
\details{
The Holme-Kim network model is a simple extension of B-A model. It adds an additional step, called "Triad formation", with the probability \emph{pt} that compensates the low clustering in B-A networks.
}
\examples{
{generate_holme_kim (1000, 20, 0.1)}
}
\references{
Holme, Petter, and Beom Jun Kim. "Growing scale-free networks with tunable clustering."Physical review E65, no. 2 (2002): 026107.
}
\author{
Xu Dong, Nazrul Shaikh
}
