% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_random}
\alias{generate_random}
\title{Generates a Erdos Renyi random graph}
\usage{
generate_random(
  size,
  probability = 0.2,
  directed = FALSE,
  allow_cycles = FALSE
)
}
\arguments{
\item{size}{is the number of nodes}

\item{probability}{is the probability of edge formation between nodes}

\item{directed}{generates directed graph when TRUE. Default value is FALSE}

\item{allow_cycles}{produces loops in the graph when TRUE. Default value is FALSE}
}
\value{
igraph object
}
\description{
Generates a Erdos Renyi random graph
}
