% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_ring}
\alias{generate_ring}
\title{Generates a ring-structured graph, in which nodes are connected with neighbours within given distance}
\usage{
generate_ring(size, distance)
}
\arguments{
\item{size}{is the number of nodes}

\item{distance}{defines maximum distance each node has to its farthest direct neighbour}
}
\value{
igraph object
}
\description{
Generates a ring-structured graph, in which nodes are connected with neighbours within given distance
}
