% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_scale_free}
\alias{generate_scale_free}
\title{Generates a Barabasi scale-free graph}
\usage{
generate_scale_free(
  size,
  preference = 1,
  directed = FALSE,
  allow_cycles = FALSE
)
}
\arguments{
\item{size}{is the number of nodes}

\item{preference}{is the power of preferencial attachment. Default is linear, i.e. 1}

\item{directed}{generates directed graph when TRUE. Default value is FALSE}

\item{allow_cycles}{produces loops in the graph when TRUE. Default value is FALSE}
}
\value{
igraph object
}
\description{
Generates a Barabasi scale-free graph
}
