% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_small_world}
\alias{generate_small_world}
\title{Generates a Watts & Strogatz small-world graph by rewiring a random graph, while keeping the degree distribution consistent}
\usage{
generate_small_world(
  size,
  probability = 0.1,
  directed = FALSE,
  allow_cycles = FALSE
)
}
\arguments{
\item{size}{is the number of nodes}

\item{probability}{is the probability of edge formation between nodes}

\item{directed}{generates directed graph when TRUE. Default value is FALSE}

\item{allow_cycles}{produces loops in the graph when TRUE. Default value is FALSE}
}
\value{
igraph object
}
\description{
Generates a Watts & Strogatz small-world graph by rewiring a random graph, while keeping the degree distribution consistent
}
