% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{generate_tree}
\alias{generate_tree}
\title{Generates a tree-structured graph}
\usage{
generate_tree(size, children = 2, direction = "undirected")
}
\arguments{
\item{size}{is the number of nodes}

\item{children}{is the number of children each node has (in addition to a parent node)}

\item{direction}{defines whether the edges are directed inwards, outwards or undirected. Possible values can be 'in', 'out' and 'undirected' (default)}
}
\value{
igraph object
}
\description{
Generates a tree-structured graph
}
