% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{get_adaptive_rank}
\alias{get_adaptive_rank}
\title{Returns ranks from 1 to highest rank before the graph is discontinued, using scores of nodes (e.g. Degree, Pagerank, Coreness, etc.)}
\usage{
get_adaptive_rank(
  graph,
  ranking_method = c("CORENESS", "PAGERANK", "COLLECTIVE_INFLUENCE", "DEGREE",
    "ECCENTRICITY", "AVERAGE_DISTANCE", "BARYCENTER", "BETWEENNESS", "BOTTLENECK",
    "CENTROID", "CLOSENESS", "CLUSTERRANK", "COMMUNITY_BETWEENNESS",
    "COMMUNITY_CENTRALITY", "CROSS_CLIQUE", "CURRENTFLOW_CLOSENESS", "DECAY",
    "EDGE_PERCOLATION", "EIGENVECTOR", "ENTROPY", "FREEMAN_CLOSENESS", "GEODESIC_K_PATH",
    "HUBBELL", "KATZ", "LAPLACIAN", "LATORA_CLOSENESS", "LEADERRANK", "LEVERAGE",
    "LINCENT", "LOBBY", "MARKOV", "MAX_NEIGHBORHOOD_COMPONENT",     
    "MAX_NEIGHBORHOOD_DENSITY", "PAIRWISE_DISCONNECTIVITY", "RADIALITY",
    "RESIDUAL_CLOSENESS", "SALSA", "SEMILOCAL", "TOPOLOGICAL_COEFFICIENT",
    "VITALITY_CLOSENESS"),
  mode = "all"
)
}
\arguments{
\item{graph}{the igraph object}

\item{ranking_method}{the adaptive method to use. Value must be "DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR", "ECCENTRICITY", "CORENESS", "PAGERANK", "COLLECTIVE_INFLUENCE"}
}
\value{
vector of ranks
}
\description{
Returns ranks from 1 to highest rank before the graph is discontinued, using scores of nodes (e.g. Degree, Pagerank, Coreness, etc.)
}
