% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{get_adaptive_ranking}
\alias{get_adaptive_ranking}
\title{Returns ranks from 1 to highest rank before the graph is discontinued, for nodes in given graph using adaptive variation of given method}
\usage{
get_adaptive_ranking(
  g,
  method = c("degree", "closeness", "betweenness", "coreness", "eigenvector",
    "eccentricity", "pagerank", "collective_influence")
)
}
\arguments{
\item{g}{the igraph object}

\item{method}{the adaptive method to use. Allowed values are degree, closeness, betweenness, coreness, eigenvector, eccentricity, pagerank, collective_influence}
}
\value{
vector of ranks
}
\description{
Returns ranks from 1 to highest rank before the graph is discontinued, for nodes in given graph using adaptive variation of given method
}
