% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{get_graph_traits}
\alias{get_graph_traits}
\title{Calculates several traits from given graph and returns as data frame}
\usage{
get_graph_traits(
  graph,
  normalize = FALSE,
  node_traits = c("degree", "betweenness", "closeness", "eigenvalue", "eccentricity",
    "coreness", "pagerank", "ci", "a-degree", "a-betweenness", "a-closeness",
    "a-eigenvalue", "a-coreness", "a-pagerank", "a-ci"),
  graph_traits = c("graph_size", "graph_edges", "graph_avg_degree", "graph_max_degree",
    "graph_apl", "graph_clust_coef", "graph_diameter", "graph_density",
    "graph_assortativity", "graph_avg_distance", "graph_triads", "graph_girth"),
  verbose = FALSE
)
}
\arguments{
\item{graph}{is the igraph object}

\item{normalize}{uses pnorm function to normalize the traits. Default is FALSE}

\item{node_traits}{is the vector of several node metrices}

\item{graph_traits}{is the vector of several graph metrices}

\item{verbose}{prints the progress log on console when TRUE. Default is FALSE}
}
\value{
data frame containing graph and its traits
}
\description{
Calculates several traits from given graph and returns as data frame
}
