% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{get_influence}
\alias{get_influence}
\title{Quantifies the influence of a set of nodes in a graph}
\usage{
get_influence(
  graph,
  nodes,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{nodes}{the set of nodes to calculate influence for}

\item{test_method}{specifies the method to measure influence. Value "RESILIENCE" (number of total nodes REMOVED (NOT THE REMAINING ones as in original resilience function) from the graph); "INFLUENCE_IC" (see simulate_ic method); "INFLUENCE_LT" (see simulate_lt method). Default is "RESILIENCE"}
}
\value{
vector of resiliences of provided combinations
}
\description{
Quantifies the influence of a set of nodes in a graph
}
\examples{
{
graph <- erdos.renyi.game(100, 0.2)
get_influence(graph=graph, nodes=V(graph)[1:10], test_method="RESILIENCE")
}
}
