% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{get_node_influence_traits}
\alias{get_node_influence_traits}
\title{Calculates several node influence traits from given graph and returns as a list}
\usage{
get_node_influence_traits(
  graph,
  normalize = FALSE,
  traits = c("betweenness", "closeness", "eigenvalue", "coreness", "pagerank", "ci",
    "a-degree", "a-betweenness", "a-closeness", "a-eigenvalue", "a-coreness",
    "a-pagerank", "a-ci"),
  verbose = FALSE
)
}
\arguments{
\item{graph}{is the igraph object}

\item{normalize}{uses pnorm function to normalize the traits. Default is FALSE}

\item{traits}{is the vector of several influential metrices (traits). Available metrices are: betweenness, closeness, eigenvalue, coreness, pagerank, ci, a-degree, a-betweenness, a-closeness, a-eigenvalue, a-coreness, a-pagerank, a-ci (a-xxx representing adaptive ranking variant)}

\item{verbose}{prints the progress log on console when TRUE. Default is FALSE}
}
\value{
data frame containing graph and its traits
}
\description{
Calculates several node influence traits from given graph and returns as a list
}
