% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_util.R
\name{get_prediction_results}
\alias{get_prediction_results}
\title{Returns Accuracy/Precision/Recall/F1-Score table}
\usage{
get_prediction_results(x, y, positive)
}
\arguments{
\item{x}{vector of values computed by classification model}

\item{y}{vector of actual values to compare with \code{x}}

\item{positive}{character value (class) to be considered as positive in \code{x} and \code{y}}
}
\value{
values of prediction metrics
}
\description{
Returns Accuracy/Precision/Recall/F1-Score table
}
\examples{
{
get_prediction_results(c(1, 1, 0, 0, 1, 1, 1), c(1, 1, 0, 0, 1, 0, 1), '1')
}
}
