% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{graph_summary}
\alias{graph_summary}
\title{Gives a summary of common metrics of given graph}
\usage{
graph_summary(graph, plot = FALSE)
}
\arguments{
\item{graph}{is the igraph object}

\item{plot}{uses tkplot to plot the \code{graph}. Default is FALSE}
}
\value{
object containing summary
}
\description{
Gives a summary of common metrics of given graph
}
\examples{
{
graph_summary(erdos.renyi.game(20, 0.2))
}
}
