% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{greedy_influential}
\alias{greedy_influential}
\title{Implements Greedy algorithm for Influence Maximization}
\usage{
greedy_influential(graph, budget, steps, prob = 0.5, test_method)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{steps}{is the time steps for which, the diffusion process should run. If exhaustive run is required, provide a high value (like 100). Default value is 1}

\item{prob}{is the probability of activation of a neighbour node. Default is 0.5}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}
}
\value{
output containing summary
}
\description{
Implements Greedy algorithm for Influence Maximization
}
\examples{
{greedy_influential(erdos.renyi.game(500, 0.005), budget=5, steps=99, prob=0.5, "RESILIENCE")}
}
