% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{h_index}
\alias{h_index}
\title{Returns H-index (Hirsch number)}
\usage{
h_index(graph, node_id)
}
\arguments{
\item{graph}{the igraph object}

\item{node_id}{is the ID of the target node}
}
\value{
h-index of the target node in the graph
}
\description{
Returns H-index (Hirsch number)
}
\examples{
{h_index(graph=erdos.renyi.game(100, 0.2), 1)}
}
