% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{ic_spread}
\alias{ic_spread}
\title{Calculates spread under IC model}
\usage{
ic_spread(graph, seed, runs = 100)
}
\arguments{
\item{graph}{is the weighted igraph object}

\item{seed}{is a set of seed (initial nodes)}

\item{runs}{is the number of times the loop should run}
}
\value{
output average spread
}
\description{
Calculates spread under IC model
}
\examples{
{
graph <- erdos.renyi.game(500, 0.005)
ic_spread(graph, seed=c(2,5,9,23), runs=10)
}
}
