% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{influence}
\alias{influence}
\title{Mine the most influential nodes in given graph}
\usage{
influence(
  graph,
  budget = 1,
  prob = 0.5,
  steps = 1,
  optimal_solution = FALSE,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC"),
  heuristic = c("GREEDY", "PAGERANK", "COLLECTIVE_INFLUENCE", "CORENESS", "CENTRALITY",
    "ADAPTIVE_CENTRALITY"),
  centrality_method = c("DEGREE", "ECCENTRICITY", "AVERAGE_DISTANCE", "BARYCENTER",
    "BETWEENNESS", "BOTTLENECK", "CENTROID", "CLOSENESS", "CLUSTERRANK",
    "COMMUNITY_BETWEENNESS", "COMMUNITY_CENTRALITY", "CROSS_CLIQUE",
    "CURRENTFLOW_CLOSENESS", "DECAY", "EDGE_PERCOLATION", "EIGENVECTOR", "ENTROPY",
    "FREEMAN_CLOSENESS", "GEODESIC_K_PATH", "HUBBELL", "KATZ", "LAPLACIAN",
    "LATORA_CLOSENESS", "LEADERRANK", "LEVERAGE", "LINCENT", "LOBBY", "MARKOV",
    "MAX_NEIGHBORHOOD_COMPONENT", "MAX_NEIGHBORHOOD_DENSITY", "PAIRWISE_DISCONNECTIVITY",
         "RADIALITY", "RESIDUAL_CLOSENESS", "SALSA", "SEMILOCAL",
    "TOPOLOGICAL_COEFFICIENT", "VITALITY_CLOSENESS"),
  parallel = TRUE,
  logging = TRUE
)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{prob}{probability at which a node influences its neighbours}

\item{steps}{is the time steps for which, the diffusion process should run. Provide NULL for exhaustive run. Default value is 1}

\item{optimal_solution}{should be TRUE if influential nodes are to be derived using optimal algorithm. Caution! This is the slowest apporach}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}

\item{heuristic}{specifies the heuristic method used for influence calculation. Required only when optimal_solution is FALSE}

\item{centrality_method}{is the centrality algorithm to use when heuristic is "CENTRALITY" or "ADAPTIVE_CENTRALITY". Value must be "DEGREE", "BETWEENNESS", "CLOSENESS" or "EIGENVECTOR"}

\item{parallel}{when true, executes the funtion using multiple CPU cores. Default value is TRUE}

\item{logging}{when true, a complete log is stored in output.log file}
}
\value{
object containing: 1. Vector of influential nodes. 2. Measure of influence. 3. Elapsed time in seconds.
}
\description{
Mine the most influential nodes in given graph
}
