% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{influence_ic}
\alias{influence_ic}
\title{Calculates influence of k nodes under Independent Cascade model}
\usage{
influence_ic(graph, seed, steps, prob)
}
\arguments{
\item{graph}{is the igraph object}

\item{seed}{is the initial seed nodes passed}

\item{steps}{is the time steps for which, the diffusion process should run. If exhaustive run is required, provide a high value (like 100). Default value is 1}

\item{prob}{is the probability of activation of a neighbour node. This is applicable only to IC model currently}
}
\value{
output containing summary, including no. of nodes activated and time taken
}
\description{
Calculates influence of k nodes under Independent Cascade model
}
