% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{influence_lt}
\alias{influence_lt}
\title{Calculates influence of k nodes under Linear Threshold model}
\usage{
influence_lt(graph, seed, steps, threshold)
}
\arguments{
\item{graph}{is the igraph object}

\item{seed}{is the initial seed nodes passed}

\item{steps}{is the time steps for which, the diffusion process should run. If exhaustive run is required, provide a high value (like 100). Default value is 1}

\item{threshold}{is minimum threshold required to activate a node under observation}
}
\value{
output containing summary, including no. of nodes activated and time taken
}
\description{
Calculates influence of k nodes under Linear Threshold model
}
