% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{katz_centrality}
\alias{katz_centrality}
\title{Calculates Katz centrality (Katz Status Index) of network. Original implementation has been modified to escape the invalid alpha value error which happens when a graph is not acyclic}
\usage{
katz_centrality(graph, alpha = 0.1)
}
\arguments{
\item{graph}{The input graph as igraph object}

\item{alpha}{The alpha parameter, which must be between 0.0-0.2. The default is 0.1.}

\item{vids}{Vertex sequence, the vertices for which the centrality values are returned. Default is all vertices.}
}
\value{
A numeric vector contaning the centrality scores for the selected vertices.
}
\description{
Katz centrality computes the relative influence of a node within a network by measuring the number of the immediate neighbors (first degree nodes) and also all other nodes in the network that connect to the node under consideration through these immediate neighbors
}
\examples{
g <- barabasi.game(20)
katz_centrality(g)
}
\references{
Junker, Bjorn H., Dirk Koschutzki, and Falk Schreiber. "Exploration of biological network centralities with CentiBiN." BMC bioinformatics 7.1 (2006): 219.
}
\author{
Mahdi Jalili \email{m_jalili@farabi.tums.ac.ir}
}
