% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{large_scale_community}
\alias{large_scale_community}
\title{Detects communities in large-scale graphs}
\usage{
large_scale_community(graph, mode = "all")
}
\arguments{
\item{graph}{is igraph object}

\item{mode}{is detection mode. Values can be "all", "in" or "out". Default is "all"}
}
\description{
Detects communities in large-scale graphs
}
\references{
Raghavan, Usha Nandini, Réka Albert, and Soundar Kumara. "Near linear time algorithm to detect community structures in large-scale networks." Physical Review E 76.3 (2007): 036106.
}
