% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_arxiv_collaboration_graph}
\alias{load_arxiv_collaboration_graph}
\title{A collaboration graph crawled from arXiv.org, High Energy Physics – Theory, from year 1991 to year 2003}
\usage{
load_arxiv_collaboration_graph()
}
\value{
igraph object
}
\description{
A collaboration graph crawled from arXiv.org, High Energy Physics – Theory, from year 1991 to year 2003
}
