% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_as_caida_graph}
\alias{load_as_caida_graph}
\title{The undirected network of autonomous systems of the Internet connected with each other from the CAIDA project, collected in 2007}
\usage{
load_as_caida_graph()
}
\value{
igraph object
}
\description{
The undirected network of autonomous systems of the Internet connected with each other from the CAIDA project, collected in 2007
}
