% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_author_netscience_graph}
\alias{load_author_netscience_graph}
\title{Coauthorship network of scientists working on network theory and experiment, as compiled by M. Newman in May, 2006}
\usage{
load_author_netscience_graph()
}
\value{
igraph object
}
\description{
Coauthorship network of scientists working on network theory and experiment, as compiled by M. Newman in May, 2006
}
