% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_football_graph}
\alias{load_football_graph}
\title{Network of American football games between Division IA colleges during regular season Fall 2000, as compiled by M. Girvan and M. Newman}
\usage{
load_football_graph()
}
\value{
igraph object
}
\description{
Network of American football games between Division IA colleges during regular season Fall 2000, as compiled by M. Girvan and M. Newman
}
