% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_influence_citation_graph}
\alias{load_influence_citation_graph}
\title{Citation network of papers on influence mining from 1990 to 2015}
\usage{
load_influence_citation_graph()
}
\value{
igraph object
}
\description{
Citation network of papers on influence mining from 1990 to 2015
}
