% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_ita2000_graph}
\alias{load_ita2000_graph}
\title{Air Transport Network is an undirected graph where nodes represent airports and edges represent a direct flight from one airport to the other}
\usage{
load_ita2000_graph()
}
\value{
igraph object
}
\description{
Air Transport Network is an undirected graph where nodes represent airports and edges represent a direct flight from one airport to the other
}
