% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_jdk6_dependencies_graph}
\alias{load_jdk6_dependencies_graph}
\title{Class dependency network of the Java Development Kit 1.6.0.7 framework}
\usage{
load_jdk6_dependencies_graph()
}
\value{
igraph object
}
\description{
Class dependency network of the Java Development Kit 1.6.0.7 framework
}
