% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_karate_club_graph}
\alias{load_karate_club_graph}
\title{A social network of members of a karate club at an American University. The network represents ties between members of the club based on their social interactions both within and away from the club}
\usage{
load_karate_club_graph()
}
\value{
igraph object
}
\description{
A social network of members of a karate club at an American University. The network represents ties between members of the club based on their social interactions both within and away from the club
}
