% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_nematode_neural_network_graph}
\alias{load_nematode_neural_network_graph}
\title{Neural network of the nematode C. Elegans compiled by Duncan Watts and Steven Strogatz from original experimental data by White et al.}
\usage{
load_nematode_neural_network_graph()
}
\value{
igraph object
}
\description{
Neural network of the nematode C. Elegans compiled by Duncan Watts and Steven Strogatz from original experimental data by White et al.
}
