% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_political_blog_graph}
\alias{load_political_blog_graph}
\title{A directed network of hyperlinks between weblogs on US politics, recorded in 2005 by Adamic and Glance}
\usage{
load_political_blog_graph()
}
\value{
igraph object
}
\description{
A directed network of hyperlinks between weblogs on US politics, recorded in 2005 by Adamic and Glance
}
