% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_protein_barabasi_graph}
\alias{load_protein_barabasi_graph}
\title{Protein interaction network data for yeast (by Barabasi)}
\usage{
load_protein_barabasi_graph()
}
\value{
igraph object
}
\description{
Protein interaction network data for yeast (by Barabasi)
}
