% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_twitter_graph}
\alias{load_twitter_graph}
\title{Network of twitter links collected via snowball sampling, started from root account seekme_94}
\usage{
load_twitter_graph()
}
\value{
igraph object
}
\description{
Network of twitter links collected via snowball sampling, started from root account seekme_94
}
