% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_wordnet_graph}
\alias{load_wordnet_graph}
\title{Lexical network of words from the WordNet dataset. Nodes in the network are English words, and links are relationships between them, such as synonymy, antonymy, meronymy, etc.}
\usage{
load_wordnet_graph()
}
\value{
igraph object
}
\description{
Lexical network of words from the WordNet dataset. Nodes in the network are English words, and links are relationships between them, such as synonymy, antonymy, meronymy, etc.
}
