% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{load_world_trade_graph}
\alias{load_world_trade_graph}
\title{Network of trade between manufacturers from around the world}
\usage{
load_world_trade_graph()
}
\value{
igraph object
}
\description{
Network of trade between manufacturers from around the world
}
