% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{lt_spread}
\alias{lt_spread}
\title{Calculates spread under LT model}
\usage{
lt_spread(graph, seed, runs = 100)
}
\arguments{
\item{graph}{is the weighted igraph object}

\item{seed}{is a set of seed (initial nodes)}

\item{runs}{is the number of times the loop should run}
}
\value{
output average spread
}
\description{
Calculates spread under LT model
}
