% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{normalize_trait}
\alias{normalize_trait}
\title{Normalizes the numeric values passed in \code{x} between 0 and 1}
\usage{
normalize_trait(x)
}
\arguments{
\item{x}{is data to be normalized}
}
\value{
normalized \code{x}
}
\description{
Normalizes the numeric values passed in \code{x} between 0 and 1
}
