% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{optimal_influential}
\alias{optimal_influential}
\title{Implements optimal algorithm for Influence Maximization}
\usage{
optimal_influential(
  graph,
  budget,
  prob = 0.5,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC"),
  parallel = TRUE
)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{prob}{probability at which a node influences its neighbours. In case of INFLUENCE_LT, this is becomes the threshold value. Default is 0.5}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}

\item{parallel}{when true, executes the funtion using multiple CPU cores. Default value is TRUE}
}
\value{
object containing: 1. Vector of influential nodes. 2. Measure of influence. 3. Elapsed time in seconds.
}
\description{
Implements optimal algorithm for Influence Maximization
}
\references{
Kempe, D., Kleinberg, J., & Tardos, É. (2003). Maximizing the Spread of Influence through a Social Network. Proceedings of the Ninth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining - KDD ’03, 137.
}
