% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{pagerank_influential}
\alias{pagerank_influential}
\title{Returns the most influential nodes in a graph using Pagerank heuristic}
\usage{
pagerank_influential(
  graph,
  budget = 1,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}
}
\value{
object containing: 1. Vector of influential nodes. 2. Measure of influence. 3. Elapsed time in seconds.
}
\description{
Returns the most influential nodes in a graph using Pagerank heuristic
}
\references{
Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The PageRank Citation Ranking: Bringing Order to the Web.
}
