% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{plot_degree_distribution}
\alias{plot_degree_distribution}
\title{Plots degree distribution of given graph}
\usage{
plot_degree_distribution(graph)
}
\arguments{
\item{graph}{is the igraph object}
}
\description{
Plots degree distribution of given graph
}
