% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{resilience}
\alias{resilience}
\title{Calculates resilience of network}
\usage{
resilience(graph, nodes)
}
\arguments{
\item{graph}{is the weighted igraph object}

\item{nodes}{is a set of nodes to check resilience of}
}
\value{
number of remaining nodes in largest connected component after removing nodes
}
\description{
Calculates resilience of network
}
\examples{
{
graph <- erdos.renyi.game(500, 0.005)
resilience(graph, nodes=V(graph)[c(2,5,9,23)])
}
}
