% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{select_adaptive_seed}
\alias{select_adaptive_seed}
\title{This function inputs an igraph object and returns budget\% nodes as seed using the adaptive method provided}
\usage{
select_adaptive_seed(
  graph,
  budget,
  seed_method = c("a-degree", "a-closeness", "a-betweenness", "a-coreness",
    "a-eigenvector")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{defines what percentage of most influential nodes out of all nodes is required as output. Default value is 1}

\item{seed_method}{is the selection method for seed (initial nodes). Value can be a-degree", "a-closeness", "a-betweenness", "a-coreness", "a-eigenvector"}
}
\value{
seed object containing set of vertices
}
\description{
This function inputs an igraph object and returns budget\% nodes as seed using the adaptive method provided
}
