% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{setup}
\alias{setup}
\title{Setup logging}
\usage{
setup(logging = TRUE)
}
\arguments{
\item{logging}{flag to enable loggging. Default is TRUE}
}
\description{
This function should be called before any other
}
