% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{simulate_ic}
\alias{simulate_ic}
\title{Simulates influence spread under Independent Cascade model}
\usage{
simulate_ic(graph, active)
}
\arguments{
\item{graph}{is the weighted igraph object}

\item{active}{represents number of active nodes in the graph}
}
\value{
number of nodes activated during simulation
}
\description{
Simulates influence spread under Independent Cascade model
}
