% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{simulate_lt}
\alias{simulate_lt}
\title{Simulates influence spread under Linear Threshold model}
\usage{
simulate_lt(graph, active, threshold = 0.5)
}
\arguments{
\item{graph}{is the weighted igraph object}

\item{active}{represents number of active nodes in the graph}

\item{threshold}{is the linear threshold between 0 and 1 with which a node is influenced}
}
\value{
number of nodes activated during simulation
}
\description{
Simulates influence spread under Linear Threshold model
}
