% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{centrality_influential}
\alias{centrality_influential}
\title{Returns the most influential nodes in a graph using centrality-based heuristics}
\usage{
centrality_influential(
  graph,
  budget = 1,
  test_method = c("RESILIENCE", "INFLUENCE_LT", "INFLUENCE_IC"),
  centrality_method = c("DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}

\item{centrality_method}{defines the centrality method to be used. Value must be:}
}
\value{
object containing: 1. Vector or influential nodes. 2. Measure of influence. 3. Elapsed time in seconds.
}
\description{
Returns the most influential nodes in a graph using centrality-based heuristics
}
