% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{get_adaptive_rank}
\alias{get_adaptive_rank}
\title{Returns ranks from 1 to highest rank before the graph is discontinued, using scores of nodes (e.g. Degree, Pagerank, Coreness, etc.)}
\usage{
get_adaptive_rank(
  graph,
  ranking_method = c("DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR",
    "ECCENTRICITY", "CORENESS", "PAGERANK", "COLLECTIVE_INFLUENCE")
)
}
\arguments{
\item{graph}{the igraph object}

\item{ranking_method}{the adaptive method to use. Value must be "DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR", "ECCENTRICITY", "CORENESS", "PAGERANK", "COLLECTIVE_INFLUENCE"}
}
\value{
vector of ranks
}
\description{
Returns ranks from 1 to highest rank before the graph is discontinued, using scores of nodes (e.g. Degree, Pagerank, Coreness, etc.)
}
