% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{get_centrality_scores}
\alias{get_centrality_scores}
\title{Returns the centrality values of nodes in a graph using given method}
\usage{
get_centrality_scores(
  graph,
  centrality_method = c("DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR")
)
}
\arguments{
\item{graph}{is the igraph object}

\item{centrality_method}{defines the centrality method to be used. Value must be:}
}
\value{
vector of centrality scores
}
\description{
Returns the centrality values of nodes in a graph using given method
}
\examples{
{
graph <- erdos.renyi.game(500, 0.05)
get_centrality_scores(graph, centrality_method=c"DEGREE")
get_centrality_scores(graph, centrality_method=c"BETWEENNESS")
get_centrality_scores(graph, centrality_method=c"CLOSENESS")
get_centrality_scores(graph, centrality_method=c"EIGENVECTOR")
}
}
