% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{get_graph_traits}
\alias{get_graph_traits}
\title{Calculates several traits from given graph and returns as data frame}
\usage{
get_graph_traits(
  graph,
  normalize = FALSE,
  graph_traits = c("SIZE", "EDGES", "AVERAGE_DEGREE", "MAX_DEGREE",
    "AVERAGE_PATH_LENGTH", "CLUSTERING_COEFFICIENT", "DIAMETER", "DENSITY",
    "ASSORTATIVITY", "AVERAGE_DISTANCE", "TRIADS", "GIRTH"),
  node_traits = c("AVERAGE_DISTANCE", "BARYCENTER", "BETWEENNESS", "BOTTLENECK",
    "CENTROID", "CLOSENESS", "CLUSTERRANK", "COMMUNITY_BETWEENNESS",
    "COMMUNITY_CENTRALITY", "CROSS_CLIQUE", "CURRENTFLOW_CLOSENESS", "DECAY",
    "EDGE_PERCOLATION", "EIGENVECTOR", "ENTROPY", "FREEMAN_CLOSENESS", "GEODESIC_K_PATH",
    "HUBBELL", "KATZ", "LAPLACIAN", "LATORA_CLOSENESS", "LEADERRANK", "LEVERAGE",
    "LINCENT", "LOBBY", "MARKOV", "MAX_NEIGHBORHOOD_COMPONENT",
    "MAX_NEIGHBORHOOD_DENSITY", "PAIRWISE_DISCONNECTIVITY", "RADIALITY",
    "RESIDUAL_CLOSENESS",      "SALSA", "SEMILOCAL", "TOPOLOGICAL_COEFFICIENT",
    "VITALITY_CLOSENESS", "CORENESS", "PAGERANK", "COLLECTIVE_INFLUENCE",
    "ADAPTIVE_AVERAGE_DISTANCE", "ADAPTIVE_BARYCENTER", "ADAPTIVE_BETWEENNESS",
    "ADAPTIVE_BOTTLENECK", "ADAPTIVE_CENTROID", "ADAPTIVE_CLOSENESS",
    "ADAPTIVE_CLUSTERRANK", "ADAPTIVE_COMMUNITY_BETWEENNESS", "COMMUNITY_CENTRALITY",
    "ADAPTIVE_CROSS_CLIQUE", "ADAPTIVE_CURRENTFLOW_CLOSENESS", "ADAPTIVE_DECAY",
    "ADAPTIVE_EDGE_PERCOLATION", "ADAPTIVE_EIGENVECTOR", "ADAPTIVE_ENTROPY",
    "FREEMAN_CLOSENESS",      "ADAPTIVE_GEODESIC_K_PATH", "ADAPTIVE_HUBBELL",
    "ADAPTIVE_KATZ", "ADAPTIVE_LAPLACIAN", "ADAPTIVE_LATORA_CLOSENESS",
    "ADAPTIVE_LEADERRANK", "ADAPTIVE_LEVERAGE", "ADAPTIVE_LINCENT", "LOBBY",
    "ADAPTIVE_MARKOV", "ADAPTIVE_MAX_NEIGHBORHOOD_COMPONENT",
    "ADAPTIVE_MAX_NEIGHBORHOOD_DENSITY", "ADAPTIVE_PAIRWISE_DISCONNECTIVITY",
    "ADAPTIVE_RADIALITY", "RESIDUAL_CLOSENESS", "ADAPTIVE_SALSA", "ADAPTIVE_SEMILOCAL",
    "ADAPTIVE_TOPOLOGICAL_COEFFICIENT", "ADAPTIVE_VITALITY_CLOSENESS",
    "ADAPTIVE_CORENESS", "ADAPTIVE_PAGERANK",      "ADAPTIVE_COLLECTIVE_INFLUENCE"),
  mode = "all",
  verbose = FALSE
)
}
\arguments{
\item{graph}{is the igraph object}

\item{normalize}{uses pnorm function to normalize the traits. Default is FALSE}

\item{graph_traits}{is the vector of several graph metrices to calculate}

\item{node_traits}{is the vector of several node metrices to calculate}

\item{verbose}{prints the progress log on console when TRUE. Default is FALSE}
}
\value{
data frame containing graph and its traits
}
\description{
Calculates several traits from given graph and returns as data frame
}
