% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{get_graph_traits}
\alias{get_graph_traits}
\title{Calculates several traits from given graph and returns as data frame}
\usage{
get_graph_traits(
  graph,
  normalize = FALSE,
  graph_traits = c("SIZE", "EDGES", "AVERAGE_DEGREE", "MAX_DEGREE",
    "AVERAGE_PATH_LENGTH", "CLUSTERING_COEFFICIENT", "DIAMETER", "DENSITY",
    "ASSORTATIVITY", "AVERAGE_DISTANCE", "TRIADS", "GIRTH"),
  node_traits = c("DEGREE", "BETWEENNESS", "CLOSENESS", "EIGENVECTOR", "ECCENTRICITY",
    "CORENESS", "PAGERANK", "COLLECTIVE_INFLUENCE", "ADAPTIVE_DEGREE",
    "ADAPTIVE_BETWEENNESS", "ADAPTIVE_CLOSENESS", "ADAPTIVE_EIGENVALUE",
    "ADAPTIVE_ECCENTRICITY", "ADAPTIVE_CORENESS", "ADAPTIVE_PAGERANK",
    "ADAPTIVE_COLLECTIVE_INFLUENCE"),
  verbose = FALSE
)
}
\arguments{
\item{graph}{is the igraph object}

\item{normalize}{uses pnorm function to normalize the traits. Default is FALSE}

\item{graph_traits}{is the vector of several graph metrices to calculate}

\item{node_traits}{is the vector of several node metrices to calculate}

\item{verbose}{prints the progress log on console when TRUE. Default is FALSE}
}
\value{
data frame containing graph and its traits
}
\description{
Calculates several traits from given graph and returns as data frame
}
