% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_maximization.R
\name{greedy_influential}
\alias{greedy_influential}
\title{Implements Greedy algorithm for Influence Maximization}
\usage{
greedy_influential(graph, budget, prob = 0.5, test_method)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{number of influential nodes to be fetched. Default value is 1}

\item{prob}{probability at which a node influences its neighbours. In case of INFLUENCE_LT, this is becomes the threshold value. Default is 0.5}

\item{test_method}{specifies the method to measure influence. Value MUST be "RESILIENCE", "INFLUENCE_IC" or "INFLUENCE_LT"}
}
\value{
output containing summary
}
\description{
Implements Greedy algorithm for Influence Maximization
}
\examples{
{greedy_influential(erdos.renyi.game(500, 0.005), budget=5, prob=0.5, "RESILIENCE")}
}
\references{
Kempe, D., Kleinberg, J., & Tardos, É. (2003). Maximizing the Spread of Influence through a Social Network. Proceedings of the Ninth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining - KDD ’03, 137.
}
